migrates source nav
